backend/requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary==2.*
requests==2.25.0
pandas==1.*
sshtunnel==0.4.0
django-environ==0.8.1
cryptography==36.0.2
PyJWT==1.7.1
django-cors-headers==3.12.0
gunicorn==20.1.0
whitenoise==6.2.0
